refactor argparse::args into submodules
